requirements-dev.txt
-r requirements.txt

flake8
flake8-commas
flake8-import-order
flake8-bugbear
flake8-builtins
pep8-naming

mypy
pandas-stubs

pytest
pytest-cov

jupyter
